ai/ai_requirements.txt
flask==2.0.1
python-chess==1.0.0
numpy==1.21.2
gunicorn==20.1.0